requirements.txt
Flask
flask-cors
tensorflow
Pillow
numpy
gdown
waitress
